Remove duplicated proxy e2e test
