--- test/features/cloudnative/proxy/proxy.go
@@ -1,109 +1,0 @@
-//go:build e2e
-
-package cloudnativeproxy
-
-import (
-	"context"
-	"testing"
-
-	dynatracev1beta1 "github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta1/dynakube"
-	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubeobjects/env"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/components/dynakube"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/kubeobjects/daemonset"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/kubeobjects/namespace"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/proxy"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/sample"
-	"github.com/Dynatrace/dynatrace-operator/test/helpers/tenant"
-	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/e2e-framework/pkg/envconf"
-	"sigs.k8s.io/e2e-framework/pkg/features"
-)
-
-const (
-	httpsProxy = "https_proxy"
-)
-
-// Prerequisites: istio service mesh
-//
-// Setup: CloudNative deployment with CSI driver
-//
-// Verification that the operator and all deployed OneAgents are able to communicate
-// over a http proxy.
-//
-// Connectivity in the dynatrace namespace and sample application namespace is restricted to
-// the local cluster. Sample application is installed. The test checks if DT_PROXY environment
-// variable is defined in the *dynakube-oneagent* container and the *application container*.
-func Feature(t *testing.T, proxySpec *dynatracev1beta1.DynaKubeProxy) features.Feature {
-	builder := features.New("cloudnative with proxy")
-	builder.WithLabel("name", "cloudnative-proxy")
-	secretConfig := tenant.GetSingleTenantSecret(t)
-
-	testDynakube := *dynakube.New(
-		dynakube.WithApiUrl(secretConfig.ApiUrl),
-		dynakube.WithCloudNativeSpec(&dynatracev1beta1.CloudNativeFullStackSpec{}),
-		dynakube.WithProxy(proxySpec),
-	)
-
-	sampleNamespace := *namespace.New("proxy-sample", namespace.WithIstio())
-	sampleApp := sample.NewApp(t, &testDynakube,
-		sample.AsDeployment(),
-		sample.WithNamespace(sampleNamespace),
-	)
-
-	// Register sample namespace create and delete
-	builder.Assess("create sample namespace", namespace.Create(sampleNamespace))
-	builder.Teardown(sampleApp.Uninstall())
-
-	// Register proxy create and delete
-	proxy.SetupProxyWithTeardown(t, builder, testDynakube)
-	proxy.CutOffDynatraceNamespace(builder, proxySpec)
-	proxy.IsDynatraceNamespaceCutOff(builder, testDynakube)
-
-	// Register dynakube install
-	dynakube.Install(builder, helpers.LevelAssess, &secretConfig, testDynakube)
-	// Register sample app install
-	builder.Assess("install sample app", sampleApp.Install())
-
-	// Register actual test
-	builder.Assess("check env variables of oneagent pods", checkOneAgentEnvVars(testDynakube))
-	builder.Assess("check proxy settings in ruxitagentproc.conf", proxy.CheckRuxitAgentProcFileHasProxySetting(*sampleApp, proxySpec))
-
-	// Register operator and dynakube uninstall
-	dynakube.Delete(builder, helpers.LevelTeardown, testDynakube)
-	return builder.Feature()
-}
-
-func checkOneAgentEnvVars(dynakube dynatracev1beta1.DynaKube) features.Func {
-	return func(ctx context.Context, t *testing.T, envConfig *envconf.Config) context.Context {
-		resources := envConfig.Client().Resources()
-		err := daemonset.NewQuery(ctx, resources, client.ObjectKey{
-			Name:      dynakube.OneAgentDaemonsetName(),
-			Namespace: dynakube.Namespace,
-		}).ForEachPod(func(podItem corev1.Pod) {
-			require.NotNil(t, podItem)
-			require.NotNil(t, podItem.Spec)
-
-			checkEnvVarsInContainer(t, podItem, dynakube.OneAgentDaemonsetName(), httpsProxy)
-		})
-
-		require.NoError(t, err)
-		return ctx
-	}
-}
-
-func checkEnvVarsInContainer(t *testing.T, podItem corev1.Pod, containerName string, envVar string) {
-	for _, container := range podItem.Spec.Containers {
-		if container.Name == containerName {
-			require.NotNil(t, container.Env)
-			require.True(t, env.IsIn(container.Env, envVar))
-			for _, env := range container.Env {
-				if env.Name == envVar {
-					require.NotNil(t, env.Value)
-				}
-			}
-		}
-	}
-}
